+++ src/commands/hardis/project/generate/gitdelta.ts
@@ -4,16 +4,8 @@ import { Messages } from "@salesforce/core";
 import { AnyJson } from "@salesforce/ts-types";
 import * as c from "chalk";
 import * as path from "path";
-import {
-  createTempDir,
-  ensureGitRepository,
-  execCommand,
-  getGitRepoRoot,
-  git,
-  gitCheckOutRemote,
-  selectGitBranch,
-  uxLog,
-} from "../../../../common/utils";
+import { createTempDir, ensureGitRepository, git, gitCheckOutRemote, selectGitBranch, uxLog } from "../../../../common/utils";
+import { callSfdxGitDelta } from "../../../../common/utils/gitUtils";
 import { prompts } from "../../../../common/utils/prompts";
 
 // Initialize Messages with the current plugin directory
@@ -125,13 +117,7 @@ export default class GenerateGitDelta extends SfdxCommand {
 
     // Generate package.xml & destructiveChanges.xml using sfdx git delta
     const tmpDir = await createTempDir();
-    const packageXmlCommand = `sfdx sgd:source:delta` + ` --from ${fromCommit}` + ` --to ${toCommit} --output ${tmpDir}`;
-    await execCommand(packageXmlCommand, this, {
-      output: true,
-      fail: true,
-      debug: this.debugMode,
-      cwd: await getGitRepoRoot(),
-    });
+    await callSfdxGitDelta(fromCommit, toCommit, tmpDir, { debug: this.debugMode });
 
     const diffPackageXml = path.join(tmpDir, "package", "package.xml");
     const diffDestructiveChangesXml = path.join(tmpDir, "destructiveChanges", "destructiveChanges.xml");