+++ src/common/gitProvider/github.ts
@@ -99,8 +99,9 @@ export class GithubProvider extends GitProviderRoot {
         repo: this.repoName,
         pull_number: prNumber,
       });
+      // Add cross git provider properties used by sfdx-hardis
       if (pullRequest) {
-        return pullRequest.data;
+        return this.completePullRequestInfo(pullRequest.data);
       }
     }
     // Case when we find PRs from a commit
@@ -151,7 +152,7 @@ export class GithubProvider extends GitProviderRoot {
         (pr: any) => pr.node.merged === true && pr.node.baseRef.name === currentGitBranch,
       );
       if (candidatePullRequests.length > 0) {
-        return candidatePullRequests[0].node;
+        return this.completePullRequestInfo(candidatePullRequests[0].node);
       }
     }
     uxLog(this, c.grey(`[GitHub Integration] Unable to find related Pull Request Info`));
@@ -228,5 +229,12 @@ _Provided by [sfdx-hardis](https://sfdx-hardis.cloudity.com) from job [${githubW
       };
       return prResult;
     }
+  }
+
+  private completePullRequestInfo(prData: any) {
+    const prInfo: any = Object.assign({}, prData);
+    prInfo.sourceBranch = (prData?.head?.ref || "").replace("refs/heads/", "");
+    prInfo.targetBranch = (prData?.base?.ref || "").replace("refs/heads/", "");
+    return prInfo;
   }
 }